Tweak for NixOS compatibility.
bash is not POSIX, but /usr/bin/env is, and NixOS can find bash this
way. Please double-check that it also works for you.

Also, on at least NixOS and Debian, the default GIDs overlap with the
container's GIDs from Ubuntu. Here, the `-f` flag avoids a spurious
failure. I was otherwise able to get `./start.sh` to work perfectly.